requirements.txt
pymupdf
pdfminer.six
pdf2image
pytesseract
pyyaml
